requirements.txt
amqp==5.3.1; python_version >= '3.6'
asgiref==3.8.1; python_version >= '3.8'
async-timeout==5.0.1; python_version >= '3.8'
billiard==4.2.1; python_version >= '3.7'
celery==5.5.2; python_version >= '3.8'
certifi==2025.4.26; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.2; python_version >= '3.7'
click==8.2.0; python_version >= '3.10'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
cryptography==44.0.3; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-rest-auth[with_social]==7.0.1; python_version >= '3.8'
django==5.2.1; python_version >= '3.10'
django-allauth[socialaccount]==65.8.0; python_version >= '3.8'
django-cors-headers==4.7.0; python_version >= '3.9'
django-debug-toolbar==5.2.0; python_version >= '3.9'
django-filter==25.1; python_version >= '3.9'
djangorestframework==3.16.0; python_version >= '3.9'
djangorestframework-simplejwt==5.5.0; python_version >= '3.9'
djoser==2.3.1; python_version >= '3.8' and python_version < '4.0'
drf-yasg==1.21.10; python_version >= '3.6'
idna==3.10; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
kombu==5.5.3; python_version >= '3.8'
oauthlib==3.2.2; python_version >= '3.6'
packaging==25.0; python_version >= '3.8'
prompt-toolkit==3.0.51; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyjwt[crypto]==2.9.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.0; python_version >= '3.9'
python3-openid==3.2.0
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
redis==6.0.0; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
social-auth-app-django==5.4.3; python_version >= '3.9'
social-auth-core==4.6.1; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
typing-extensions==4.13.2; python_version >= '3.8'
tzdata==2025.2; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.4.0; python_version >= '3.9'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13